Re ordeno las clases y endpoint para el cliente.

+++ src/main/java/com/soporte/repository/TicketRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface TicketRepository extends CrudRepository<Ticket, Long>{ }
+public interface TicketRepository extends CrudRepository<Ticket, Integer>{ }
